+++ modules/Attendance/Percent.php
@@ -216,7 +216,7 @@ if ( ! $_REQUEST['modfunc'] )
 			'AVERAGE_ABSENT' => round( issetVal( $sum['AVERAGE_ABSENT'], 0 ), 1 ),
 		];
 
-		// Force display of $link['add'] on PDF or if not allowed to edit
+		// Force display of $link['add'] on PDF, Export or if not allowed to edit
 		$options['add'] = true;
 
 		ListOutput(