+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 12.6
 ---------------
 - HTML responsive stackable tables in Export.php
+- Export list: add Total row in ListOutput.fnc.php
 
 Changes in 12.5
 ---------------